requirements.txt
# standard 
pytest
pytest-cov
pylint
black
ipython
# standard py app specific
setuptools
# app specific 
requests
python-dateutil
# my packages
git+https://github.com/dexplorer/utils#egg=utils
git+https://github.com/dexplorer/df-metadata#egg=metadata
